fix(Like) : merge로 인한 충돌 해결

+++ src/main/java/org/noostak/likes/application/LikeService.java
@@ -4,5 +4,5 @@ import org.noostak.likes.dto.IncreaseResponse;
 
 public interface LikeService {
     IncreaseResponse increase(Long memberId, Long appointmentId, Long appointmentOptionId);
-    int getLikeCountByOptionId(Long optionId);
+    Long getLikeCountByOptionId(Long optionId);
 }
